backend/requirements.txt
Django
djangorestframework
django-cors-headers
graphene-django
django-filter
psycopg2-binary
python-decouple
